refactor: dto를 반환하는 레포지토리 엔티티를 반환하는 레포지토리 분리
