+++ backend/ddang/src/main/java/com/ddang/ddang/chat/application/ChatRoomService.java
@@ -13,6 +13,7 @@ import com.ddang.ddang.chat.application.exception.ChatRoomNotFoundException;
 import com.ddang.ddang.chat.application.exception.InvalidAuctionToChatException;
 import com.ddang.ddang.chat.application.exception.UserCannotAccessChatRoomException;
 import com.ddang.ddang.chat.domain.ChatRoom;
+import com.ddang.ddang.chat.infrastructure.persistence.ChatRoomAndMessageRepository;
 import com.ddang.ddang.chat.infrastructure.persistence.JpaChatRoomRepository;
 import com.ddang.ddang.chat.infrastructure.persistence.dto.ChatRoomAndMessageDto;
 import com.ddang.ddang.user.application.exception.UserNotFoundException;
@@ -33,6 +34,7 @@ public class ChatRoomService {
     private static final Long DEFAULT_CHAT_ROOM_ID = null;
 
     private final JpaChatRoomRepository chatRoomRepository;
+    private final ChatRoomAndMessageRepository chatRoomAndMessageRepository;
     private final JpaUserRepository userRepository;
     private final JpaAuctionRepository auctionRepository;
 
@@ -82,7 +84,7 @@ public class ChatRoomService {
         final User findUser = userRepository.findById(userId)
                                             .orElseThrow(() -> new UserNotFoundException("사용자 정보를 찾을 수 없습니다."));
         final List<ChatRoomAndMessageDto> chatRoomAndMessageQueryProjectionDtos =
-                chatRoomRepository.findAllChatRoomInfoByUserIdOrderByLastMessage(findUser.getId());
+                chatRoomAndMessageRepository.findAllChatRoomInfoByUserIdOrderByLastMessage(findUser.getId());
 
         return chatRoomAndMessageQueryProjectionDtos.stream()
                                                     .map(dto -> ReadChatRoomWithLastMessageDto.of(findUser, dto))
